+++ test-apps/express-v4/src/app.d.ts
@@ -1,7 +1,0 @@
-/// <reference types="lucia-auth" />
-declare namespace Lucia {
-	type Auth = import('./lucia.js').Auth;
-	type UserAttributes = {
-		username: string;
-	};
-}